pacify the merciless clippy

+++ src/input/input_field.rs
@@ -91,7 +91,7 @@ where
     }
 
     fn fmt_index(&self, index: Option<crate::Id>, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
-        fmt_index(C::FIELD_DEBUG_NAMES[self.field_index as usize], index, fmt)
+        fmt_index(C::FIELD_DEBUG_NAMES[self.field_index], index, fmt)
     }
 }
